Fixed interpreter to try multiple evolvers until it finds one that works.
In the course of this added an additional test: This has flagged up that
XText is not currently managing the Henshin registration correctly (it
seems to actually remove this from the package registry between test
runs). As a result, whichever of the two interpreter tests gets run
second will fail because it cannot find the Henshin meta-model. Need to
look into how to set up XText correctly.